Move callMonitor to fritzFon module.
